+++ py/desitarget/test/test_mtl.py
@@ -14,8 +14,8 @@ class TestMTL(unittest.TestCase):
         self.types = np.array(['ELG', 'LRG', 'QSO', 'QSO', 'ELG'])
         self.priorities = [Mx[t].priorities['UNOBS'] for t in self.types]
         self.post_prio = [Mx[t].priorities['MORE_ZGOOD'] for t in self.types]
-        self.post_prio[0] = 1  #- ELG
-        self.post_prio[2] = 1  #- low-z QSO
+        self.post_prio[0] = 2 #ELG
+        self.post_prio[2] = 2 #lowz QSO
         self.targets['DESI_TARGET'] = [Mx[t].mask for t in self.types]
         self.targets['BGS_TARGET'] = np.zeros(len(self.types), dtype=np.int64)
         self.targets['MWS_TARGET'] = np.zeros(len(self.types), dtype=np.int64)
@@ -49,8 +49,9 @@ class TestMTL(unittest.TestCase):
     def test_zcat(self):
         mtl = make_mtl(self.targets, self.zcat, trim=False)
         mtl.sort(keys='TARGETID')
-        self.assertTrue(np.all(mtl['NUMOBS_MORE'] == [0, 1, 0, 3, 1]))
         self.assertTrue(np.all(mtl['PRIORITY'] == self.post_prio))
+        self.assertTrue(np.all(mtl['NUMOBS_MORE'] == [0, 1, 0, 3, 1]))
+
         
         #- change one target to a SAFE (BADSKY) target and confirm priority=0 not 1
         self.targets['DESI_TARGET'][0] = Mx.BADSKY